pxe: refactor pxe and dhcp_vm containers
Merge pxe and dhcp_vm containers together. The new container can work as
DHCP server for the VM or as a PXE server.

With the separation of Jenkins pipelines, the container build step is done
in the Jenkins_build pipelines which does not have access to Ansible
inventory information. So we can't customize the Jenkins container with
docker buildargs anymore. The customization must now be done with
environment variables during the docker run. To facilitate this we add a
script that will take care of this step.

The PXE server can work with BIOS and UEFI.

Issue: #6633
Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>
Change-Id: Icf9039e1cac9691f94764c6dcdec9836f63fc056